refactor(examples/gamepad-snes): share start(canvas) between browser and GJS
Follow the canvas2d-fireworks showcase pattern: extract a single
start(canvas) entry point in snes-gamepad-demo.ts that both platforms
call identically. Browser and GJS entries are now <15 lines each.

Also:
- Add LT/RT trigger buttons to SNES controller SVG + BUTTON_MAP
- Both platforms now render triggers (visible behind shoulder buttons)
